LayoutTests/imported/w3c/web-platform-tests/intersection-observer/same-document-zero-size-target-expected.txt


PASS Observing a zero-area target. 
PASS First rAF 
FAIL document.scrollingElement.scrollTop = 300 assert_equals: entries.length expected 2 but got 1
FAIL document.scrollingElement.scrollTop = 100 assert_equals: entries.length expected 3 but got 1
